requirements.txt
PyMuPDF>=1.24.0
ezdxf>=1.3.0
Pillow>=10.0.0
reportlab>=4.0.0
PyYAML>=6.0
pytest>=8.0.0
